+++ server/initialization/index.js
@@ -3,7 +3,6 @@ let debug = require("debug")("app:initialization");
 const INIT_SECTIONS = [
     require("./sections/helmet"),
     require("./sections/body_parsing"),
-    require("./sections/route_rewriting"),
     require("./sections/static_content"),
     require("./sections/sessions"),
     require("./sections/passport"),
