--- server/initialization/sections/route_rewriting.js
@@ -1,12 +1,0 @@
-"use strict";
-
-module.exports = {
-    name: "route rewriting",
-    configure(app) {
-        app.get(/^(?!\/api|.*\.).*/, (req, res, next) => {
-            req.url = "/index.html";
-            next();
-        });
-    }
-};
-
